refactor: streamline component imports and update prop documentation
