+++ src/components/index.js
@@ -1,0 +1,6 @@
+export { default as CustomButton } from './CustomButton'
+export { default as CustomCard } from './CustomCard'
+export { default as CustomFooter } from './CustomFooter'
+export { default as CustomInteractive } from './CustomInteractive'
+export { default as CustomNavBar } from './CustomNavBar'
+export { default as FAQComponent } from './FAQComponent'
