+++ src/screens/Error404.jsx
@@ -1,8 +1,7 @@
 
 const Error404 = () => {
   return (
-    <div className="flex flex-col flex-1 items-center justify-center"
-    >
+    <div className='flex flex-col flex-1 items-center justify-center'>
       <h1>404 Page not found</h1>
       <p>The page you are looking for does not exist.</p>
     </div>
