+++ src/screens/Session.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const Session = () => {
   return (
