+++ src/screens/Settings.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const Settings = () => {
   return (
